link_bio/requirements.txt
reflex==0.4.5
python-dotenv
supabase
